python-container/requirements.txt
websocket-client==0.58.0
confluent-kafka==2.3.0